Component migration: Start second phase by renaming `products` folder to `components` (#1258)
