+++ src/components/breadcrumbs/template.test.tsx
@@ -1,6 +1,6 @@
 import React from 'react'
 import { render, screen } from '@testing-library/react'
-import Breadcrumbs from '.'
+import Breadcrumbs from './template'
 
 describe('<Breadcrumbs />', () => {
   const defaultProps = {
