+++ src/components/heading/formatted-type.ts
